送信API追加。JSONのStringを作るための関数を追加。Checkboxの所持属性、メソッドを追加。詳しい処理はコード見てね:munou:

+++ Heatbuster/app/src/main/java/io/github/joken/heatbuster/TemperatureStatus.java
@@ -23,7 +23,7 @@ public enum TemperatureStatus {
 		this.brand = br;
 	}
 
-	public static TemperatureStatus getStatusbyTemp(int temp){
+	public static TemperatureStatus getStatusbyTemp(float temp){
 		if(temp <= Safe_Limen){
 			return Safe;
 		}else{
